+++ client/src/state.rs
@@ -1,7 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod bucket;
 pub mod client;
 pub mod key_store;
 pub mod snapshot;
