+++ engine/vault/src/types.rs
@@ -5,7 +5,6 @@ use std::mem;
 
 pub use transactions::{DataTransaction, SealedBlob, SealedTransaction};
 
-pub mod ntransactions;
 pub mod transactions;
 pub mod utils;
