+++ engine/vault/tests/vault.rs
@@ -4,327 +4,139 @@
 mod utils;
 use utils::provider::Provider;
 
-use stronghold_utils::test_utils::fresh;
-
-use vault::{DBView, Key, Kind, PreparedRead, ReadResult, RecordHint, RecordId, Result, WriteRequest};
-
-use std::{collections::HashMap, iter::empty};
-
-use rand::Rng;
-
-fn write_to_read(wr: &WriteRequest) -> ReadResult {
-    ReadResult::new(wr.kind(), &wr.id(), &wr.data())
-}
-
-fn fresh_record_hint() -> RecordHint {
-    let mut bs = [0; 24];
-    rand::thread_rng().fill(&mut bs);
-    bs.into()
-}
+use vault::{DbView, Key, RecordHint, RecordId, VaultId};
 
 #[test]
-fn test_empty() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v = DBView::load(k, empty::<ReadResult>())?;
-
-    assert_eq!(v.all().len(), 0);
-    assert_eq!(v.absolute_balance(), (0, 0));
-    assert_eq!(v.chain_ctrs(), HashMap::new());
-    assert_eq!(v.gc().len(), 0);
+fn test_vaults() {
+    let mut view: DbView<Provider> = DbView::new();
 
-    Ok(())
-}
+    let key0 = Key::random().unwrap();
+    let vid0 = VaultId::random::<Provider>().unwrap();
+    let rid0 = RecordId::random::<Provider>().unwrap();
+    let rid01 = RecordId::random::<Provider>().unwrap();
 
-#[test]
-fn test_truncate() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
+    let key1 = Key::random().unwrap();
+    let vid1 = VaultId::random::<Provider>().unwrap();
+    let rid1 = RecordId::random::<Provider>().unwrap();
 
-    let mut writes = vec![];
+    // init two vaults.
+    view.init_vault(&key0, vid0).unwrap();
+    view.init_vault(&key1, vid1).unwrap();
+    // write to vault0 and record0
+    view.write(&key0, vid0, rid0, b"test0", RecordHint::new(b"hint").unwrap())
+        .unwrap();
+    // write to vault0 and record01
+    view.write(&key0, vid0, rid01, b"test01", RecordHint::new(b"hint").unwrap())
+        .unwrap();
+    // write to vault1 and record1
+    view.write(&key1, vid1, rid1, b"test1", RecordHint::new(b"hint").unwrap())
+        .unwrap();
 
-    let id = RecordId::random::<Provider>()?;
-    writes.push(v0.writer(id).truncate()?);
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
-
-    assert_eq!(v1.all().len(), 1);
-    assert_eq!(v1.absolute_balance(), (1, 1));
-    assert_eq!(v1.chain_ctrs(), vec![(id, 0u64)].into_iter().collect());
-    assert_eq!(v1.gc().len(), 0);
-
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::RecordIsEmpty);
-
-    Ok(())
-}
-
-#[test]
-fn test_read_non_existent_record() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v = DBView::load(k, empty::<ReadResult>())?;
-
-    let id = RecordId::random::<Provider>()?;
-    assert_eq!(v.reader().prepare_read(&id)?, PreparedRead::NoSuchRecord);
-
-    Ok(())
-}
-
-#[test]
-fn test_write_cache_hit() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
+    let list0 = view.list_hints_and_ids(&key0, vid0);
 
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    let data = fresh::bytestring();
-    let hint = fresh_record_hint();
-    writes.append(&mut w.write(&data, hint)?);
+    assert_eq!(list0.len(), 2);
 
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
+    // read from vault0 and record0
+    view.get_guard(&key0, vid0, rid0, |g| {
+        assert_eq!(b"test0", &(*g.borrow()));
 
-    assert_eq!(v1.all().len(), 1);
-    assert_eq!(v1.records().count(), 1);
-    assert_eq!(v1.absolute_balance(), (2, 2));
-    assert_eq!(v1.chain_ctrs(), vec![(id, 1u64)].into_iter().collect());
-    assert_eq!(v1.gc().len(), 0);
+        Ok(())
+    })
+    .unwrap();
 
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::CacheHit(data));
+    // read from vault0 and record01
+    view.get_guard(&key0, vid0, rid01, |g| {
+        assert_eq!(b"test01", &(*g.borrow()));
 
-    Ok(())
-}
+        Ok(())
+    })
+    .unwrap();
 
-#[test]
-fn test_write_cache_miss() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
+    // read from vault1 and record1
+    view.get_guard(&key1, vid1, rid1, |g| {
+        assert_eq!(b"test1", &(*g.borrow()));
 
-    let mut writes = vec![];
+        Ok(())
+    })
+    .unwrap();
 
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    let data = fresh::bytestring();
-    let hint = fresh_record_hint();
-    let (bid, blob) = match w.write(&data, hint)?.as_slice() {
-        [w0, w1] => {
-            assert_eq!(w0.kind(), Kind::Transaction);
-            writes.push(w0.clone());
+    // revoke records rid01 and rid1
+    view.revoke_record(&key0, vid0, rid01).unwrap();
+    view.revoke_record(&key1, vid1, rid1).unwrap();
 
-            assert_eq!(w1.kind(), Kind::Blob);
-            (w1.id().to_vec(), w1.data().to_vec())
-        }
-        ws => panic!("{} unexpected writes", ws.len()),
-    };
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
+    let list0 = view.list_hints_and_ids(&key0, vid0);
 
-    let r = v1.reader();
-    let res = match r.prepare_read(&id)? {
-        PreparedRead::CacheMiss(req) => {
-            assert_eq!(req.id(), bid.as_slice());
-            req.result(blob)
-        }
-        x => panic!("unexpected value: {:?}", x),
-    };
-
-    assert_eq!(r.read(res)?, data);
-
-    Ok(())
-}
-
-#[test]
-fn test_write_twice() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
-
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    let data0 = fresh::bytestring();
-    let data1 = fresh::bytestring();
-    let hint = fresh_record_hint();
-    writes.append(&mut w.write(&data0, hint)?);
-    writes.append(&mut w.write(&data1, hint)?);
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
-
-    assert_eq!(v1.all().len(), 1);
-    assert_eq!(v1.records().count(), 1);
-    assert_eq!(v1.absolute_balance(), (2, 3));
-    assert_eq!(v1.chain_ctrs(), vec![(id, 2u64)].into_iter().collect());
-    assert_eq!(v1.gc().len(), 1);
-
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::CacheHit(data1));
-
-    Ok(())
-}
-
-#[test]
-fn test_revoke() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
-
-    let id = RecordId::random::<Provider>()?;
-    writes.push(v0.writer(id).truncate()?);
-
-    let v1 = DBView::load(k.clone(), writes.iter().map(write_to_read))?;
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::RecordIsEmpty);
-    writes.push(v1.writer(id).revoke()?);
-
-    let v2 = DBView::load(k, writes.iter().map(write_to_read))?;
-    assert_eq!(v2.reader().prepare_read(&id)?, PreparedRead::NoSuchRecord);
-
-    assert_eq!(v2.all().len(), 1);
-    assert_eq!(v2.records().count(), 0);
-    assert_eq!(v2.absolute_balance(), (0, 2));
-    assert_eq!(v2.chain_ctrs(), vec![(id, 1u64)].into_iter().collect());
-    assert_eq!(v2.gc().len(), 2);
-
-    Ok(())
-}
-
-#[test]
-fn test_rekove_without_reload() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
-
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    writes.push(w.revoke()?);
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::NoSuchRecord);
-
-    assert_eq!(v1.all().len(), 1);
-    assert_eq!(v1.records().count(), 0);
-    assert_eq!(v1.absolute_balance(), (0, 2));
-    assert_eq!(v1.chain_ctrs(), vec![(id, 1u64)].into_iter().collect());
-    assert_eq!(v1.gc().len(), 2);
-
-    Ok(())
-}
-
-#[test]
-fn test_rekove_then_write() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
-
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    writes.push(w.revoke()?);
-    let data = fresh::bytestring();
-    let hint = fresh_record_hint();
-    writes.append(&mut w.write(&data, hint)?);
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
-    assert_eq!(v1.reader().prepare_read(&id)?, PreparedRead::CacheHit(data));
-
-    assert_eq!(v1.all().len(), 1);
-    assert_eq!(v1.records().count(), 1);
-    assert_eq!(v1.absolute_balance(), (2, 3));
-    assert_eq!(v1.chain_ctrs(), vec![(id, 2u64)].into_iter().collect());
-    assert_eq!(v1.gc().len(), 1);
-
-    Ok(())
-}
-
-// #[test]
-// #[ignore = "not yet implemented: we need some kind of checksum in the data transaction to protect against this case:
-// when the users key is compromised"] fn test_ensure_authenticty_of_blob() -> Result<()> {
-//     let k: Key<Provider> = Key::random()?;
-//     let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-//     let mut writes = vec![];
-
-//     let id = RecordId::random::<Provider>()?;
-//     let mut w = v0.writer(id);
-//     writes.push(w.truncate()?);
-//     let hint = fresh_record_hint();
-//     let bid = match w.write(&fresh::bytestring(), hint)?.as_slice() {
-//         [w0, w1] => {
-//             assert_eq!(w0.kind(), Kind::Transaction);
-//             writes.push(w0.clone());
-
-//             assert_eq!(w1.kind(), Kind::Blob);
-//             w1.id().to_vec()
-//         }
-//         ws => panic!("{} unexpected writes", ws.len()),
-//     };
-
-//     let v1 = DBView::load(k.clone(), writes.iter().map(write_to_read))?;
-
-//     let r = v1.reader();
-//     let res = match r.prepare_read(&id)? {
-//         PreparedRead::CacheMiss(req) => {
-//             let data: SealedBlob = fresh::bytestring().encrypt(&k, bid)?;
-//             req.result(data.as_ref().to_vec())
-//         }
-//         x => panic!("unexpected value: {:?}", x),
-//     };
-
-//     match r.read(res) {
-//         Err(vault::Error::ProtocolError(_)) => (),
-//         Err(_) | Ok(_) => panic!("unexpected result"),
-//     }
-
-//     Ok(())
-// }
-
-#[test]
-fn test_storage_returns_stale_blob() -> Result<()> {
-    let k: Key<Provider> = Key::random()?;
-    let v0 = DBView::load(k.clone(), empty::<ReadResult>())?;
-
-    let mut writes = vec![];
-
-    let id = RecordId::random::<Provider>()?;
-    let mut w = v0.writer(id);
-    writes.push(w.truncate()?);
-    let hint = fresh_record_hint();
-
-    let (bid, blob) = match w.write(&fresh::bytestring(), hint)?.as_slice() {
-        [w0, w1] => {
-            assert_eq!(w0.kind(), Kind::Transaction);
-            assert_eq!(w1.kind(), Kind::Blob);
-            (w1.id().to_vec(), w1.data().to_vec())
-        }
-        ws => panic!("{} unexpected writes", ws.len()),
-    };
-
-    match w.write(&fresh::bytestring(), hint)?.as_slice() {
-        [w0, w1] => {
-            assert_eq!(w0.kind(), Kind::Transaction);
-            writes.push(w0.clone());
-
-            assert_eq!(w1.kind(), Kind::Blob);
-        }
-        ws => panic!("{} unexpected writes", ws.len()),
-    };
-
-    let v1 = DBView::load(k, writes.iter().map(write_to_read))?;
-
-    let r = v1.reader();
-    let res = match r.prepare_read(&id)? {
-        PreparedRead::CacheMiss(_) => ReadResult::new(Kind::Blob, &bid, &blob),
-        x => panic!("unexpected value: {:?}", x),
-    };
-
-    match r.read(res) {
-        Err(vault::Error::ProtocolError(_)) => (),
-        Err(_) | Ok(_) => panic!("unexpected result"),
-    }
-
-    Ok(())
+    assert_eq!(list0.len(), 1);
+
+    // garbage collect vid0.
+    view.garbage_collect_vault(&key0, vid0).unwrap();
+
+    let list0 = view.list_hints_and_ids(&key0, vid0);
+
+    assert_eq!(list0.len(), 1);
+
+    let b = view.contains_record(&key0, vid0, rid0);
+
+    assert!(b);
+
+    let b = view.contains_record(&key0, vid0, rid01);
+
+    assert!(!b);
+
+    // read vid0 and rid0.
+    view.get_guard(&key0, vid0, rid0, |g| {
+        assert_eq!(b"test0", &(*g.borrow()));
+
+        Ok(())
+    })
+    .unwrap();
+
+    let key0 = Key::random().unwrap();
+    let vid0 = VaultId::random::<Provider>().unwrap();
+    let rid0 = RecordId::random::<Provider>().unwrap();
+
+    let key1 = Key::random().unwrap();
+    let vid1 = VaultId::random::<Provider>().unwrap();
+    let rid1 = RecordId::random::<Provider>().unwrap();
+
+    // Write data into vid0/rid0
+    view.write(&key0, vid0, rid0, b"test", RecordHint::new(b"hint").unwrap())
+        .unwrap();
+
+    // execute a procedure and put the result into a new record
+    view.exec_proc(
+        &key0,
+        vid0,
+        rid0,
+        &key1,
+        vid1,
+        rid1,
+        RecordHint::new(b"tester").unwrap(),
+        |guard| {
+            let data = guard.borrow();
+            let mut ret = Vec::new();
+
+            ret.extend(data.iter());
+            ret.extend(data.iter());
+
+            Ok(ret)
+        },
+    )
+    .unwrap();
+
+    // read vid0 and rid0.
+    view.get_guard(&key0, vid0, rid0, |g| {
+        assert_eq!(b"test", &(*g.borrow()));
+
+        Ok(())
+    })
+    .unwrap();
+
+    // read vid1 and rid1.
+    view.get_guard(&key1, vid1, rid1, |g| {
+        assert_eq!(b"testtest", &(*g.borrow()));
+
+        Ok(())
+    })
+    .unwrap();
 }
